+++ 3DBook/Core/FolderAggregate/Folder.cs
@@ -6,7 +6,7 @@ namespace _3DBook.Core.FolderAggregate;
 public class Folder : EntityBase,IAggregateRoot
 {
     private Folder(){} //need for EF
-    public Folder(int folds, decimal enter, decimal exit,int machineId,string sortCode)
+    public Folder(int folds, int enter, int exit,int machineId,string sortCode)
     {
         
         GuardClauses.GuardClause.IsNegative(folds,nameof(folds));
@@ -20,7 +20,7 @@ public class Folder : EntityBase,IAggregateRoot
         MachineId = machineId;
     }
     // Add a new constructor for testing only
-    public Folder(int folds, decimal enter, decimal exit, int machineId, string sortCode, Machine machine)
+    public Folder(int folds, int enter, int exit, int machineId, string sortCode, Machine machine)
         : this(folds, enter, exit, machineId, sortCode)
     {
         Machine = machine;
@@ -28,8 +28,8 @@ public class Folder : EntityBase,IAggregateRoot
 
     public string Code {  get; private set; }
     public int Folds { get; private set; }
-    public decimal Enter { get; private set; }
-    public decimal Exit { get; private set; }
+    public int Enter { get; private set; }
+    public int Exit { get; private set; }
     public int MachineId { get; private set; }
     public Machine Machine { get; private set; }
 }
